Creating Accounts app with resume uploading, user profile, and registering a user

+++ backend/job/views.py
@@ -1,8 +1,9 @@
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework import status
 from django.db.models import Avg, Min, Count, Max
 from rest_framework.pagination import PageNumberPagination
+from rest_framework.permissions import IsAuthenticated
 
 from .serializers import JobSerilaizer
 from .models import Job
@@ -32,7 +33,9 @@ def getJob(request, pk):
 
 
 @api_view(["POST"])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data["user"] = request.user
     data = request.data
     job = Job.objects.create(**data)
     serializer = JobSerilaizer(job, many=False)
@@ -40,8 +43,13 @@ def newJob(request):
 
 
 @api_view(["PUT"])
+@permission_classes([IsAuthenticated])
 def updateJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+    if job.user != request.user:
+        return Response(
+            {"message": "You cannot update this job"}, status=status.HTTP_403_FORBIDDEN
+        )
     job.title = request.data["title"]
     job.description = request.data["description"]
     job.email = request.data["email"]
@@ -61,8 +69,13 @@ def updateJob(request, pk):
 
 
 @api_view(["DELETE"])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+    if job.user != request.user:
+        return Response(
+            {"message": "You cannot update this job"}, status=status.HTTP_403_FORBIDDEN
+        )
     job.delete()
     return Response({"message": "Job is deleted"}, status=status.HTTP_200_OK)
